requirements.txt
flask-uploads
tensorflow==1.12.0
music21
numpy==1.22.4
